Code cleanup, fix bug making the camera feed yellow-ish

+++ ESP/lib/src/data/config/project_config.hpp
@@ -44,9 +44,9 @@ class ProjectConfig : public Preferences, public ISubject<ConfigState_e> {
   struct CameraConfig_t {
     uint8_t vflip;
     uint8_t href;
-    uint8_t framesize;
+    int framesize;
     uint8_t quality;
-    uint8_t brightness;
+    int brightness;
     bool simple_auto_exposure_on;
     bool fancy_auto_exposure_on;
     int ae_level;
